fix(api): fix stop workflow run for purge
