+++ engine/api/workflow/dao_node_run.go
@@ -658,8 +658,8 @@ func updateNodeRunCommits(db gorp.SqlExecutor, id int64, commits []sdk.VCSCommit
 	return nil
 }
 
-// UpdateNodeRunStatusAndStage update just noderun status and stage
-func UpdateNodeRunStatusAndStage(db gorp.SqlExecutor, nodeRun *sdk.WorkflowNodeRun) error {
+// updateNodeRunStatusAndStage update just noderun status and stage
+func updateNodeRunStatusAndStage(db gorp.SqlExecutor, nodeRun *sdk.WorkflowNodeRun) error {
 	stagesBts, errMarshal := json.Marshal(nodeRun.Stages)
 	if errMarshal != nil {
 		return sdk.WrapError(errMarshal, "updateNodeRunStatusAndStage> Unable to marshal stages")
